Move `check_fork_next_block`
